+++ src/main/java/com/alechilles/alecstamework/persistence/diagnostics/PersistenceAutomaticDiagnosticReporter.java
@@ -1,9 +1,9 @@
 package com.alechilles.alecstamework.persistence.diagnostics;
 
-import com.alechilles.alecstelemetry.api.TelemetryDiagnosticAttachment;
-import com.alechilles.alecstelemetry.api.TelemetryDiagnosticBundle;
-import com.alechilles.alecstelemetry.api.TelemetryDiagnosticBundleResult;
-import com.alechilles.alecstelemetry.api.TelemetryDiagnosticDisposition;
+import com.alechilles.beacon.api.TelemetryDiagnosticAttachment;
+import com.alechilles.beacon.api.TelemetryDiagnosticBundle;
+import com.alechilles.beacon.api.TelemetryDiagnosticBundleResult;
+import com.alechilles.beacon.api.TelemetryDiagnosticDisposition;
 import com.alechilles.alecstamework.persistence.runtime.PersistenceFailureSignal;
 import com.hypixel.hytale.logger.HytaleLogger;